fix bugs

+++ utils/LogUtils.py
@@ -14,15 +14,18 @@ def get_month_and_day() -> Tuple[int]:
     return cur_month, cur_day
 
 
-def set_log_fold_dir(log_dir: str, log_folder: str, fold: int, is_lock: bool) -> str:
+def set_log_fold_dir(log_dir: str, log_folder: str, fold: int, is_lock: bool, use_ckpt: bool = False) -> str:
     log_fold_dir = os.path.join(log_dir, log_folder, "fold{}".format(fold))
 
-    if os.path.exists(log_fold_dir):
-        if not os.path.exists(os.path.join(log_fold_dir, "lock.txt")):
-            shutil.rmtree(log_fold_dir)
-        else:
-            raise RuntimeError("Create Log Dir failed")
-    os.makedirs(log_fold_dir)
+    if use_ckpt:
+        assert os.path.exists(log_fold_dir), "not exist dir: {}".format(log_fold_dir)
+    else:
+        if os.path.exists(log_fold_dir):
+            if not os.path.exists(os.path.join(log_fold_dir, "lock.txt")):
+                shutil.rmtree(log_fold_dir)
+            else:
+                raise RuntimeError("Create Log Dir failed")
+        os.makedirs(log_fold_dir)
 
     if is_lock:
         open(os.path.join(log_fold_dir, "lock.txt"), "w").close()
